+++ CHANGELOG.md
@@ -9,6 +9,7 @@ SECURITY
 FEATURES
 
 * Added support for publishing Docker images to Amazon ECR
+* Added support for searching and getting documentation for policies from the Terraform Registry
 * Enhanced toolset for resolving provider documentation, fetching provider docs, searching modules, and retrieving module details from the Terraform Registry.
 * Added support for Streamable HTTP, see [#99](https://github.com/hashicorp/terraform-mcp-server/pull/99)
 
@@ -35,4 +36,3 @@ FEATURES
 - Provides seamless integration with Terraform Registry APIs for provider and module discovery, documentation retrieval, and advanced IaC automation.
 - Initial support for VS Code and Claude Desktop integration.
 - Includes basic CI/CD, Docker build, and test infrastructure.
-